Fix breakage that occurred adding get for a region
Ensure we aren't resetting and getting at the same time -- which is
invalid. When adding the feature to allow getting all settings
for a specific region we removed the check to see if reset was set and
prevent also setting getConfig from being called in Run. This fixes the
issue by putting the appropriate check in place and adding tests to
ensure there is no stdout when resetting.

Fixes: lp:1634197