OpenStack firewaller now takes a clock.Clock
This change is just to speed up testing. By taking a clock.Clock deleteSecurityGroup can be tested near-instantly rather than taking 30 seconds. The provider itself now has a clock, which is initialized by default to clock.WallClock to match the existing behavior during normal operation.

QA Steps:
This is a non-change in terms of Juju's normal usage, so just running the unit tests is sufficient.