Add Prometheus monitoring to agents
We create a Prometheus registry for each
machine and unit agent, and expose it via
the agent's introspection socket. The socket
can be exposed as HTTP using socat or similar,
and then used as a regular Prometheus target.

We also add metrics for the logsender worker,
tracking the number of log records enqueued,
sent (to the controller), and dropped; and
also the capacity of the queue so we can
compute usage %.