Fix StatePool leaks
Part of http://pad.lv/1641824 - the migration back doesn't work because
the state references in the pool never get to 0, so they're marked for
removal but not actually removed.

Ensure that state references returned from
httpContext.stateForRequestUnauthenticated get released back to the
StatePool when they're finished with.

This covers the simple cases that can be handled with "defer
h.ctxt.release(st)".