Fix StatePool leaks
Part of http://pad.lv/1641824 - the migration back doesn't work because
the state references in the pool never get to 0, so they're marked for
removal but not actually removed.

Ensure that state references returned from
httpContext.stateForRequestUnauthenticated get released back to the
StatePool when they're finished with.

In the logstream handler. since the state is created in the newSource
closure and returned, the normal defer method doesn't work. Change
newSource to also return a closer function that will release the state
back to the pool and call that closer func from
logStreamRequestHandler.close().