Fix StatePool/goroutine leak in logsink handler
Part of http://pad.lv/1641824

During a migration the client disconnects and closes the socket, but the
ServeHTTP loop would stay blocked waiting because the server doesn't get
shut down so h.ctxt.stop() isn't closed. This would cause the goroutine
to be blocked indefinitely and the state to be held in the StatePool
with an outstanding reference (because the defer doesn't fire).

Close the log message channel on error from JSON.Receive to unblock the
ServeHTTP goroutine.