Release state back to the pool on errors
In the functions that do more checks after the StatePool.Get but return
the state (so can't use a deferred release because it should outlive the
scope), we need to make sure that the state is returned to the pool when
we return an error and a nil state.

Use the named return values + deferred if err != nil release trick (which
really needs a catchier name) when there's more than one error path.

Includes a drive-by fix for the charmsHandler where I was releasing the
state immediately rather than deferring it.