provider/rackspace: filter out diskless flavors
The compute and memory optimised flavors on Rackspace
are diskless, so you must boot them with a volume. For
now, we just filter those flavors out so that we won't
try to use them automatically.

Also add external-network to default config for rackspace,
the lack of which was causing bootstrap to fail.