provider/maas: persist bridge script
Previously the MAAS script that bridges interfaces was deleted from
the filesystem after it had run because it was a one-time only event.

As we are now moving towards dynamic bridging we need permanent access
to the script; the script is now persisted in /var/lib/juju. That path
is not hard-coded, it comes from calling: paths.DataDir(series).

QA steps:

1) bootstrap on MAAS
2) verify that /var/lib/juju/add-juju-bridge.py exists
3) verify that /var/lib/juju/add-juju-bridge.py is executable
4) reboot machine, and verify (2) and (3) post reboot
5) verify bridge(s) exist: $(brctl show)