core/description: Separate resource revisions
It turns out in practice that the details for the same resource revision
number can be different between the charmstore and application revision,
particularly because of local resource uploads.

Instead of storing a shared map of revisions, which are referenced by
revision number, store each revision separately. This is closer to how
they're stored in Juju's state.