Fix intermittently-failing test
I could make it reliable in by adjusting the delay, but running it with
the race checker made that break too, so I've relaxed the exact numbers
that it checks for. Really it just needs to check that there are
multiple messages being reported - the exact counts are not likely to be
wrong.

Fixes https://bugs.launchpad.net/juju/+bug/1646504