Use Model Status
This code starts us on the path to actually using status on models.
This first commit just adds code to set status, and transmit that
to juju status as needed.  The only thing to currently set status
is migration, which sets status to busy, and instead of setting
its message in a special migration field, now sets the message
in the status message.