Update Machine.SetContainerLinkLayerDevices
Instead of creating a device for every bridge that is on the Host
machine, instead we only create devices for bridges that are in the
spaces that are defined by the constraints for the container.
Further, if we cannot find a bridge for any of the spaces, then we
signal an error. (This is currently not handled properly by the
LXD broker, but we can fix that in the future.)

This does not yet attempt to create bridges for host devices that
are not bridged. That will happen in a follow up patch.

Also, if the host machine has more than one bridge in a space, then
both bridges will be used inside the container. However, if only one
of the host devices is bridged, then it will not ask to bridge the
other one.