migration: Check source controller version
As well as checking the source model version against the target
controller version, the source controller verison needs to be checked.

A migration is allowed as long as the major.minor of the target
controller is greater than or equal to the major.minor of the source
controller. This offers the possibility of migration back to the
original controller as long as the versions are close
together (i.e. have compatible migration APIs).