create some helper functions
Make it a bit easier to set up test state, we may move directly to these
helpers as it brings in the context for the test in a nice straightforward
fashion, instead of having it be implicit state somewhere else in the file.
Tweak the inference function so that we can handle the 'default' space,
and machines in just a single space. As well, make sure that we return
an error when appropriate.