Send discharged macaroons when connecting to a stream
If the macaroons passed in to an API client connection already include
discharge macaroons, the discharge request dance doesn't need to be
done, so the macaroons won't be in the cookie jar. In this case make
sure they're still sent along when making a websocket connection. (This
is analogous to the code in api.state.Login that adds macaroons from the
bakery client jar to the login request.)

Includes a drive-by fix to MacaroonSuite - the DischargerLogin function
wasn't reset in setup so the function set in one test could be called
accidentally from another test.

Fixes https://bugs.launchpad.net/juju/+bug/1650451