Add Provisioner.HostChangesForContainer()
This updates bits of state to support detecting what bridges
a machine needs to create for us to put a container onto that
machine that wants access to particular spaces.
It then exposes that via the API, and then implements the
client side of that API.

This version is targetted at the 2.1 branch, instead of
'develop'.