Remove 'default' as a magic space.
As a named space, "default" is actually a bad choice.
The issue is that it isn't the default name on MAAS, and
wouldn't be the default name anywhere else. Better to
not have a magic string at first than to confuse the issue.
Eventually want to have a default space, but named based
on how you created the model.