Fallback to a bridge-all mode.
If we find that we are on a machine where we don't understand the spaces,
and we have a container that doesn't want any spaces, we should fall back
to a bridge-all mode like we used to do.
We should still bridge late (no bridges until the first container),
and we should still handle not trying to use lxdbr0, etc.