Remove build flags, just skip tests
Had used build flags to prevent builing on windows, but we apparently
build there and just skip tests. This seems like a good place to save
some time in test runs, but not as part of this work.

This removes the build tags except immediately surrounding the bits
that won't compile on windows.

Refs: remove-uvtool-dependency