Raise all network interfaces, post bridging a single interface
In dynamic bridging we were taking down ens5.100, turning that into a
bridge, then bringing up br-ens5.100. This is not enough. We also need
to bring the physical device back up too. Rather than list the
interfaces to bring back up as "ens5.100 br-ens5.100" rely on ifup's
-a option (all interfaces marked auto) to work out what needs to be
brought back up. This divorces us from having to walk the dependencies
of VLAN's on a bond on a physical device, et al.