Don't be specific about architectures to build
Since we only support kvm on arm64, amd64, and pp64el, it seems ideal
to not biuld it elsewhere. However, why punish others where it might
just work.

Is it better to not build and explicitly state it isn't
supported on other arches or to let it build and then maybe not work?

Fixes: https://bugs.launchpad.net/bugs/1654589