provider/azure: create common resources once only
Create common resources (storage account, vnet, nsg)
once only, when creating the model. For bootstrap,
we do this along with the bootstrap machine as we
already did. For non-controller models, we create
a deployment for the common resources at model
creation time. Because deployments can take a while,
and model creation is synchronous, we do the common
resource creation asynchronously and wait for its
completion before starting any machines.