state: Add support for sequences with a minimum
Added State.sequenceWithMin() and the infrastructure to support it. This
is like State.sequence() but allows a minimum value for a sequence to be
specified. This is surprisingly complicated to implement due to there
being no atomic way to provide initial values for an upserted counter in
MongoDB.

This is preparation for supporting tracking of which local charm URLs
have been used without having to keep dead charm documents.
See https://bugs.launchpad.net/juju/+bug/1657614