start factoring out the 'ssh reachable' code.
We want to integrate with the golang crypto ssh library so that
we not only check that we can get a TCP connection to the port, but
also so that we can check that there is a valid SSH that is presenting
the right public key on the other side.
Also, our code was causing the goroutines to block indefinitely, as
they'd never be able to send on the channel once we find one that is
correct, so close a done channel to signal they have nothing to do.