Renamed UpdateController to AdoptResources
Updated the documentation to indicate that it will now be called on the
target controller (so the name makes more sense). Also this name isn't
used elsewhere so it's easier to search for.

AdoptResources now takes the version of the source controller - this can
be used to handle the tagging correctly if the technique changes across
versions.