Move builtin clouds to providers
Remove the builtin cloud definitions from
the cloud package, and add an interface
that providers can implement to "detect"
clouds. For now this is just implemented
by LXD to return the localhost cloud, but
it could also be used to detect cloud
definitions based on client-side config
such as novarc files, etc.