always record machine network config.
We were only recording it if we had provider config to merge with it.
However, this will be useful to have everywhere and it lets us move forward
with some of our explicit decisions on how we want to do bridging, etc.
Note that containers now start 'properly' on AWS, but only because
of a sequence of failures and fallbacks that gets us to invoke old
code paths.

Namely, we see the machine, we trigger the 'fallback' to bridge all
interfaces, which says 'create br-eth0'. The bridge script runs, but
fails to actually create br-eth0 because the definition of eth0 is
actually in /e/n/interfaces.d/50cloud.cfg, and it doesn't read it.
However, the call to 'prepareHost' already succeeded, so we continue
with the next step. It then asks what devices to create for the
container it is starting, which comes back with 'I dont know',
which it then says, 'well then, I'll fall back to lxdbr0', and thus
starts a container in the old way.