provider/lxd: use cloud/creds properly
We change the way the LXD provider conveys the
host address and credentials to the controller, by
using the existing environs.CloudSpec structure.

The lxd provider now reports a "localhost" cloud,
and its (new) FinalizeCloud method will fill in the
empty endpoints, equal to the address of the LXD
host on the LXD bridge interface.

An upgrade step is added that rewrites the cloud
and credential definitions in the state database.