Ensure cloud.Name is set and used everywhere
A recent change introduced the Name field to
cloud.Cloud. In several parts of the code
base we were not setting it properly. This
branch ensures that it is always filled in,
introduces additional validation in state,
and updates various bits of code to stop
passing around the cloud name separately.