be strict about binding names
When deploying an application, you're allowed to bind endpoints to spaces.
However, we would allow any endpoint name to be supplied, which meant that
typos would not be caught.
(https://pad.lv/1661929)

Further, when deploying an application and only supplying the default binding,
if that application did not have any relations at all, it wouldn't be put
into the preferred space. We now record the default binding request along
with the other endpoint bindings. (https://pad.lv/1654952)