provider/azure: handle empty authorized-keys
Azure requires that a machine have either a
password or an SSH authorised key set for a
machine. If authorized-keys is empty in the
config, generate a key pair. We will pass
the public key into Azure, and throw away
the private key.

Fixes https://bugs.launchpad.net/juju/+bug/1658830