cmd/...: fix single-key config commands
Fix controller-config and model-config commands
so that when a single key is specified, by
default we will format the value in its raw
form. Internally, we use the "smart" formatter
to do this.

To do this in a way that users can still get
YAML/JSON formatting of single keys as needed,
we introduce and default to tabular formatting
of controller config as we do for model config.

Also, fix the model-config command to honour
the -o/--output flag when a single key is
specified.

Fixes https://bugs.launchpad.net/juju/+bug/1661506