lots of refactoring, no change to functionality.
SetHostMachineNetworkConfig should be taking a names.MachineTag not an arbitrary string,
especially when the string is labeled 'ID' (eg '0') but actually contains a 'Tag' ('machine-0').
Start working on changing the lxd broker to take a 'prepareHost' instead of passing in lots of
context just so it can hand it back again.
Working towards having prepareHost grab the machine-lock while bridging.