Add InProcessUpdate to proxyupdateworker
At the moment we update environment variables with the proxy settings,
intending for them to apply to web requests made from the running
process (like charm downloads). Unfortunately the built-in
implementation that reads proxy info from the environment only reads the
variables once - if a request has already been made the variable will
have been read and subsequent changes won't be seen.

We still want the variables to be set in the environment so they'll be
right in processes that get started from this one, but for the
in-process requests we use a different default transport. This will get
the proxy settings from a package variable instead of env vars. The
function passed into the proxy updater will set this package variable.