cloduconfig: attempt agent downloads in parallel
We introduce a Python script to attempt downloading
agent binary URLs in parallel, and use it in the
cloud-init scripts run on Linux. The first URL
for which we get a response will be downloaded.

Attempting the URLs in series leads to poor agent
startup times in Azure, where the agents cannot
communicate with the controller over the private
network. The cloud-init script was prioritising the
private network when attempting URLs, and that
would cause a significant delay before the public
Internet address/URL was attempted.

Fixes https://bugs.launchpad.net/juju/+bug/1622531