Implement util.proxy.ProxyConfig
This stores default proxy settings in a package variable. The manifolds
that run the proxyupdater now pass the default config's Set function to
the worker so that it will keep the proxy config
up-to-date. InstallInDefaultTransport sets the default http transport's
proxy resolution method to use these config settings (with logic adapted
from the stdlib net/http.ProxyFromEnvironment function).