provider/lxd: improve endpoint handling
We improve the cloud endpoint handling so that
users can now specify any of the following:

 - (empty) => connect to the unix socket, identify the
   bridge device used for default profile's eth0 NIC,
   and identify the IP address assigned to the bridge
   device on the host. Query LXD for the port it is
   listening on for HTTPS requests.
 - host => connect to the specified host on the default
   port (8443) over HTTPS.
 - host:port => connect to the specified host and port
   over HTTPS.
 - https://host[:port]/path => connect to the specified
   URL over HTTPS.

In the future we should also support unix://, but for
now there are too many assumptions, so we reject any
endpoint starting with unix://.

We will no longer use the unix socket just because the
endpoint is a local address. We only use the unix socket
to create certificate credentials for the local LXD,
and to configure LXD to listen on HTTPS and obtain the
appropriate endpoint for communication from within.