lxd: update LXD for 2.9's new storage API
This is basically the same thing as the network API update commit, although
it is somewhat easier to deal with:

LXD 2.9 has a storage API (indicated by the "storage" API extension), and
now users must configure a default storage pool. This simply adds code that
detects whether there is a storage pool configured and in the default
profile. If there is, and it's in the profile, then this code is a no-op.

Otherwise, it creates a storage pool (using the directory backend; it looks
like all the ZFS code got ripped out and I assume there's a reason for
that), and attaches that storage pool to the default profile so that
containers will use it as their pool for root disks.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>