state: fix SingularSuite.TestExpire
The test code was erroneously using the test clock for
timeouts, when it should be using the system clock. The
tests would mostly succeed because the test clock was
advanced asynchronously, and the first select would
choose the second case because it was already waiting
on the (test) clock. If the clock was advanced first,
the select would block forever.

Fixes https://bugs.launchpad.net/juju/+bug/1625768