environs: drop Environ.BootstrapMessage method
The Environ.BootstrapMessage method was added
as a means of logging a message at bootstrap
time in the LXD provider. The added method is
redundant, as we already pass context to the
Bootstrap method, which the provider can use
to show messages to the user.

We drop the BootstrapMessage method, and update
the LXD provider to use the bootstrap context
to display its message.