Make user management messages more informative
This reworks the error messages when adding and removing users. It
essentially returns "user unavailable" rather than "not found" when
deleting and "is permanently deleted" rather than "user exists".

It also changes a bunch of gocheck asserts into checks where is made no
sense to fail the suite on the first error.

Refs: https://bugs.launchpad.net/juju/+bug/1630728