Handle mutex.ErrCancelled coming from worker.Stop
This happens very occasionally when stopping the uniter after a
successful test. It happens when the waitHooks step (which acquires and
releases the hook lock) blocks the uniter code from acquiring the lock -
the mutex code loops and delays for 250ms in that case, and it's still
in that delay select when the test closes the abort channel. This makes
the mutex return ErrCancelled which bubbles back up to the test.