Run actions synchronously from the command line
Improve the run-action UX by allowing the command to wait for
and return the result, rather than returning an ID. This is
by far the most common use case, and the major reason that
the run and ssh commands remain more popular than actions.
Unfortunately, we need to remain backwards compatible for
now and cannot make this the default.

A --wait option has been added to the run-action command,
which waits and returns the result. A --timeout option has
also been added, letting operators specify a timeout
rather than waiting indefinitely.