state: progress towards detachable storage
Various changes to state to progress towards
having detachable storage instances, and better
support for removing storage instances. In a
follow-up, removing the storage attachment
for a unit-owned storage instance will clear
the storage instance's "owner" field, and
decrement the unit's refcount for that storage
name. This will allow the user to attach the
storage to another unit (once attachment code
is in place).

When a storage instance is removed, we will
now destroy the volume or filesystem that is
associated with it, ignoring the "life binding".
In a follow up, we will remove the life binding
field.

When machine-bound storage is removed along
with a machine, we will also remove the storage
instance that the machine-bound storage is
associated with. Storage instances cannot be
allowed to outlive the associated machine-
level storage.

Finally: attempting to destroy a non-existent
storage instance, volume, or filesystem, will
return an error. The exception to this rule
is if the entity is removed concurrently.