Properly record machine network config before launching containers.
Added a step so that after we 'initialize' container runtimes, we then revalidate
the host's network config. This is what was broken on 'trusty' machines, because
they don't start with an lxdbr0.

We also find that Provisioner.SetObservedNetworkConfig was a copy&paste rather
than sharing the Machiner implementation, which explains why the second retry
of provisioning containers still didn't notice that we had lxdbr0.