apiserver/common/storagecommon: handle notfound
Handle not found errors when fetching the volume
or filesystem for a storage instance, when computing
storage attachment info. If the unit is not yet
assigned to a machine, the volume or filesystem
will not exist. We handle the not found error and
translate it to a "not provisioned" error, which
the layer above expects.