state: replace workers package with worker.Runner
This fixes https://bugs.launchpad.net/juju/+bug/1669540

Instead of the custom state/workers package with, we use the existing
worker.Runner implementation. This raises the issue of what to do
when a worker cannot be acquired. The obvious solution to that is
to change the worker methods to return an error, but that turns out
to be a huge change because it means that every single Watch method
needs to return an error, and hence all the callers of those.
Instead of causing massive disruption to the code base like that, we
instead observe that watchers already have a way of returning an error -
via the Wait method. So we implement a few extra functions that can
return already-dead workers in the relevant packages (state/watcher,
state/presence and worker/lease) and use those instead.