Send CIDR and VLAN tag in AddSubnets request
This will enable simplifying the discover spaces API. At the moment the
API requests these from the provider itself. It has a lot of extra
complexity for cross-checking the sent subnet information with the details
queried from the provider. This makes sense for the API underpinning
the `juju add-subnet` command, where we want the user to provide either
the provider id or the CIDR, and need to fill in the other fields. But
it's very confusing to make the discover spaces worker go through the
same path, when it has all the necessary info from the provider already.