Handle 'Node with this Hostname already exists'.
If we have a failure to actually start an LXD (or KVM) container, then
we would mark it as in error, but would retry the provisioning after a
10s delay. However, we should not try to CreateDevice() again in MAAS,
(or we should be Delete()ing the device). The way we were doing it was
causing us to Create the device twice, which was why MAAS was giving an
error the second time.

This was tested on a live MAAS by injecting a failure just before
broker.manager.CreateContainer. Which should cause us to fail to
provision the container with that specific error, rather than failing
with Hostname already exists.