state: fix mongo indexes/usage
The main issue is statuseshistory: when
pruning, we are using a raw collection,
which means model-uuid is not added to
queries automatically; and we're not
adding model-uuid in what should be a
model-specific query. Because of that,
the index that includes model-uuid is
not (cannot be) used, and instead mongo
performs a full, super slow, COLLSCAN.
The fix is to add model-uuid to the query.

The juju.txns collection lacks any indexes
at all (except the default one); this is
OK except when resuming transactions,
which performs a query on the "s" field.
I have added an index for this.