Change how we update charm bindings to handle default bindings.
The existing code was always setting all values anyway, might as well not
split it into a create+remove and just set it to an exact value.
That lets us support "" as well.

Make sure we have a test that exercises default binding.
Especially when it comes to upgrade/SetCharm testing.