apiserver/application: Destroy{Application,Unit}
Introduce new DestroyApplication and DestroyUnit
methods to the Application facade. These supersede
the existing Destroy and DestroyUnits methods,
respectively. The new methods are bulk, and return
information relating to the entities being destroyed.

This will enable us to inform the user of storage
that is destroyed, or detached and left in the model.

Also, move a few tests over to the unit test suite.