remove-machine: tell user of unit/storage changes
Introduce the MachineManager.(Force)DestroyMachines
methods, and use them in "juju remove-machine". We
will continue to use the existing Client facade
methods when the MachineManager facade is not new
enough.

Using the new methods, we print out the names of
units and storage that will be destroyed or detached
when removing the machine.