cmd/juju/user: refactor login command
We integrate the two kinds of login, so local user login
is orthogonal to whether you're logging into a public controller.

Logging into a controller is now idempotent - no message
will be printed unless the account or controller have changed.

The -u (--user) flag now takes the username as its argument,
so the argument to juju login is always now the controller domain.

Also add a few tests for existing untested behaviour and integrate
the two login test suites (but left them in two separate files for now
so that the diff looks better).