Driveby fix for shims
Stop the space- and subnetShims from embedding the interface they were
supposed to be implementing. They were never constructed with an
instance of the interface, so the effect of the embedding was to
suppress compile errors that the interface wasn't fully implemented. The
code using them happens to work because it never calls those methods,
but if it did it would compile fine and then panic at runtime. Instead
remove the unused methods (and the embedding).