Put public address in settings for c-m relations
When the current relation is a cross-model one, we use the public
address to ensure that the remote service uses an address that is
guaranteed to reach this unit.

In the future we can do more checking to determine whether the
cloud-local address can be used based on the region/tenant/etc - in that
case we can use it to short-circuit the connection. But this gives us a
good starting point that's likely to work in the most clouds, even those
with pretty restrictive network models like Azure (and maybe GCE).

In lxd there'll generally be no public address, so for now we log to
report that we're falling back to the cloud-local address. This will be
ok until we can replace it with something provider-specific - that check
will allow cloud-local address for lxd and forbid it for other clouds,
until we add more specific short-circuit cases.