state: make storage detachable/attachable
When detaching storage from a unit, we now
leave the storage in the model, and just
clear the storage instance's ownership. The
storage can then be attached to another unit.

The existing model storage cleanups have been
replaced with one that destroys storage
instances, rather than volumes and filesystems.
The machine storage will be destroyed as a
consequence.