provider/ec2: catch additional error status
When detaching a volume, if it is not attached
to any machine an IncorrectState status code
will be returned from EC2.

Also, fix an improper errors.Annotatef when
attaching volumes, which would cause a nil error
to be returned under error conditions.

Introduced a reverse proxy into the local test
server to simplify testing of edge cases.
Individual tests can manipulate the reverse
proxy to return responses that will trigger
those edge cases.